+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 * The calendar doesn't get closed an open when clicking in the input anymore.
 * Fixes errors not finding dates when customizing day rendering.
 * Event listeners in `inputProps` now only override default behaviour when returning `false`.
+* Adds the `displayTimeZone` prop. Thanks to @martingordon
 
 ## 2.15.0
 * New `onNavigateBack` and `onNavigateForward` hooks thanks to @DaanDD and @simeg.